Fix FastDDS windows CI error

+++ src/cpp/dynamic-types/DynamicDataHelper.cpp
@@ -255,7 +255,7 @@ void DynamicDataHelper::print_basic_collection(
     {
         auto count = data->get_item_count();
         std::cout << "[";
-        for (size_t i = 0; i < count; ++i)
+        for (uint32_t i = 0; i < count; ++i)
         {
             print_basic_element(data, i, data->type_->get_element_type()->get_kind());
             std::cout << (i == count - 1 ? "]" : ", ");
@@ -291,7 +291,7 @@ void DynamicDataHelper::print_complex_collection(
     {
         auto count = data->get_item_count();
 
-        for (size_t i = 0; i < count; ++i)
+        for (uint32_t i = 0; i < count; ++i)
         {
             std::cout << tabs << "[" << i << "] = ";
             print_complex_element(data, i, tabs);
